chore(workflow): update license header check
Updated the license header check in the workflow to use the correct grep command. Previously, it was using '-rcL' which caused the check to fail. Replaced it with '-rL' for proper functionality. Additionally, modified the output format to display the files with license header issues as a bullet list for better readability.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>