chore: Update linting rules and configuration
The given changes include updates to the linting rules and configuration files for the program. This includes modifications to the `.github/dependabot.yml` and `.golangci.yml` files. The changes involve the addition of new files, modifications to existing files, and the inclusion of code snippets related to package ecosystems, scheduling, assignees, reviewers, and commit messages. The configuration file applies linting rules to Go code and includes the GNU General Public License. It specifies the enabled linters and sets values for the "goheader" linter. The program also includes various features for performance and experimental purposes, such as line and statement counts, maximum complexity, duplication threshold, and enabled/disabled tags.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>