chore(ci): Update Go versions and add linting check
The commit updates the Go versions in the CI workflow from 1.19, 1.20, and 1.21 to include Go 1.18 as well. Additionally, a new linting check is added for the `ubuntu-latest` OS and Go 1.21.

This commit ensures that the CI workflow is using the correct Go versions and includes a linting check for improved code quality.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>